split large files: App.Cli.cs (1321) and CompressSettingsWindow.xaml.cs (1132)
- App.Cli.cs -> App.Compress.cs + App.Extract.cs + App.Open.cs (by command domain)
- CompressSettingsWindow.xaml.cs -> CompressSettingsWindow.Password.cs (password tab logic)
- original files kept as empty partial class shells
- 171/171 tests pass, 0 build errors
